requirements.txt
Django==2.1
wheel==0.31.1
gunicorn==19.9.0
psycopg2==2.7.5
django-ipware
